+++ Config/ProjectSettings.ps1
@@ -218,6 +218,12 @@ if (!(Get-Variable -Name CheckProjectConstants -Scope Global -ErrorAction Ignore
 	# Required minimum PowerShellGet version prior to installing other modules, do not decrement!
 	New-Variable -Name RequirePowerShellGetVersion -Scope Global -Option Constant -Value ([version]::new(2, 2, 4))
 
+	# Recommended minimum platyPS version used to generate online help files for modules, do not decrement!
+	New-Variable -Name RequirePlatyPSVersion -Scope Global -Option Constant -Value ([version]::new(0, 14, 0))
+
+	# Recommended minimum VSCode version, do not decrement!
+	New-Variable -Name RequireVSCodeVersion -Scope Global -Option Constant -Value ([version]::new(1, 48, 2))
+
 	# Recommended minimum PSReadline version for command line editing experience of PowerShell
 	# Needs the 1.6.0 or a higher version of PowerShellGet to install the latest prerelease version of PSReadLine
 	New-Variable -Name RequirePSReadlineVersion -Scope Global -Option Constant -Value ([version]::new(2, 0, 4))