+++ Rules/IPv4/Inbound/Multicast.ps1
@@ -49,7 +49,7 @@ Address Range					Size		CIDR				Designation
 . $PSScriptRoot\..\..\..\Config\ProjectSettings.ps1
 
 # Check requirements
-Initialize-Project
+Initialize-Project -Abort
 
 # Imports
 . $PSScriptRoot\DirectionSetup.ps1