+++ Rules/IPv4/Outbound/Temporary.ps1
@@ -29,7 +29,7 @@ SOFTWARE.
 . $PSScriptRoot\..\..\..\Config\ProjectSettings.ps1
 
 # Check requirements
-Initialize-Project
+Initialize-Project -Abort
 
 # Imports
 . $PSScriptRoot\DirectionSetup.ps1