+++ Rules/IPv6/Inbound/ICMP.ps1
@@ -90,7 +90,7 @@ Type    Name
 . $PSScriptRoot\..\..\..\Config\ProjectSettings.ps1
 
 # Check requirements
-Initialize-Project
+Initialize-Project -Abort
 
 # Imports
 . $PSScriptRoot\DirectionSetup.ps1