+++ Templates/Outbound.ps1
@@ -49,7 +49,7 @@ New-Variable -Name ThisScript -Scope Private -Option Constant -Value (
 	$MyInvocation.MyCommand.Name -replace ".{4}$" )
 
 # Check requirements
-Initialize-Project
+Initialize-Project -Abort
 
 # Imports
 # TODO: adjust paths