+++ Test/Project.Windows.Firewall/Remove-FirewallRules.ps1
@@ -35,7 +35,7 @@ New-Variable -Name ThisScript -Scope Private -Option Constant -Value (
 	$MyInvocation.MyCommand.Name -replace ".{4}$" )
 
 # Check requirements
-Initialize-Project
+Initialize-Project -Abort
 
 # Imports
 . $PSScriptRoot\ContextSetup.ps1