+++ Test/Project.Windows.ProgramInfo/Edit-Table.ps1
@@ -46,7 +46,7 @@ if ((Get-Variable -Name Develop -Scope Global).Value -eq $false)
 }
 
 # Check requirements
-Initialize-Project
+Initialize-Project -Abort
 
 # Imports
 . $PSScriptRoot\ContextSetup.ps1