+++ Test/Project.Windows.ProgramInfo/Find-Installation.ps1
@@ -45,7 +45,7 @@ if ((Get-Variable -Name Develop -Scope Global).Value -eq $false)
 }
 
 # Check requirements
-Initialize-Project
+Initialize-Project -Abort
 
 # Imports
 . $PSScriptRoot\ContextSetup.ps1