+++ Test/Project.Windows.ProgramInfo/Show-Table.ps1
@@ -41,7 +41,7 @@ if ((Get-Variable -Name Develop -Scope Global).Value -eq $false)
 }
 
 # Check requirements
-Initialize-Project
+Initialize-Project -Abort
 
 # Imports
 . $PSScriptRoot\ContextSetup.ps1